requirements.txt
boto3==1.28.1
botocore==1.31.1
jmespath==1.0.1
python-dateutil==2.8.2
PyYAML==6.0
s3transfer==0.6.1
six==1.16.0
urllib3==1.26.16
